Revert "Feature/frontend/refactoring"
