+++ frontend/src/components/graph/CptContainer.vue
@@ -21,7 +21,7 @@
 <script lang="ts">
 import Vue from "vue";
 import ProbTable from "@/components/graph/ProbTable.vue";
-import { dcbn } from "@/utils/graph/graph";
+import { dcbn } from "../../utils/graph";
 
 export default Vue.extend({
   components: {
