+++ frontend/src/components/graph/NodeProperties.vue
@@ -47,7 +47,7 @@
 <script lang="ts">
 import Vue from "vue";
 import CptContainer from "@/components/graph/CptContainer.vue";
-import { dcbn } from "@/utils/graph/graph";
+import { dcbn } from "../../utils/graph";
 
 export default Vue.extend({
   data() {
